v2/requirements.txt
# FinCat v2 Requirements
# Updated: December 5, 2025

# File monitoring
watchdog>=3.0.0

# Excel handling
xlrd>=2.0.1          # Read legacy .xls files
openpyxl>=3.1.2      # Read/write modern .xlsx files

# AI integration
anthropic>=0.75.0    # Claude API (UPDATED: fixed proxies issue)

# Configuration
PyYAML>=6.0.1        # config.yaml parsing
python-dotenv>=1.0.0 # .env file loading

# User experience
colorama>=0.4.6      # Colored terminal output
tqdm>=4.66.1         # Progress bars

# Testing (development only)
pytest>=7.4.0
pytest-cov>=4.1.0
